+++ benchmarks/industrial_benchmark_python/ib_eval_UDLUFT.py
@@ -113,7 +113,7 @@ class Agent_daniel_21(Ib_Agent):
         if len(self.state_history) > 10:
             self.state_history.pop()
 
-        at = np.array([0, 0, 0], dtype=np.float16)
+        at = np.array([0, 0, 0], dtype=np.float32)
         at[0] = - self.get_h('v', 5) - 0.91
         at[1] = 2 * self.get_h('f', 3) - env_state['p'] + 1.43
         at[2] = -3.48 * self.get_h('h', 3) - self.get_h('h', 4) + 2 * self.get_h('p', 0) + 0.81
@@ -131,7 +131,7 @@ class Agent_daniel_27(Ib_Agent):
         if len(self.state_history) > 10:
             self.state_history.pop()
 
-        at = np.array([0, 0, 0], dtype=np.float16)
+        at = np.array([0, 0, 0], dtype=np.float32)
         # # idk why, but these are the best??
         # at[0] = -self.get_h('v', 5) - 1.17
         # at[1] = 2 * self.get_h('f', 3) - self.get_h('p', 0) + 1.16
@@ -153,7 +153,7 @@ class Agent_Daniel_29_Best(Ib_Agent):
         if len(self.state_history) > 10:
             self.state_history.pop()
 
-        at = np.array([0, 0, 0], dtype=np.float16)
+        at = np.array([0, 0, 0], dtype=np.float32)
         at[0] = -self.get_h('v', 4) - 0.96  # 27
         at[1] = 0.41 * self.get_h('f', 1) + self.get_h('f', 4) - 0.59 * self.get_h('p', 0) + 0.77  # 27
         at[2] = -4.05 * self.get_h('h', 3) - self.get_h('h', 4) + 2.26 * self.get_h('p', 0) + 0.90  # 27
@@ -177,7 +177,7 @@ class Agent_Test(Ib_Agent):
         if len(self.state_history) > 10:
             self.state_history.pop()
 
-        at = np.array([0, 0, 0], dtype=np.float16)
+        at = np.array([0, 0, 0], dtype=np.float32)
         at[0] = -self.get_h('g', 5) - 1.8  # 27
         at[1] = 0.41 * self.get_h('f', 1) + self.get_h('f', 4) - 0.59 * self.get_h('p', 0) + 0.77  # -13.8*Consumption_30
         at[2] = -4.05 * self.get_h('h', 3) - self.get_h('h', 4) + 2.26 * self.get_h('p', 0) + 0.90  # 27
@@ -200,7 +200,7 @@ class Agent_Udluft(Ib_Agent):
         if len(self.state_history) > 10:
             self.state_history.pop()
 
-        at = np.array([0, 0, 0], dtype=np.float16)
+        at = np.array([0, 0, 0], dtype=np.float32)
         at[0] = -self.get_h('f', 0) - self.get_h('v', 3)  # 27
         at[1] = self.get_h('f', 4)
         at[2] = -6 * self.get_h('h', 4)
@@ -225,7 +225,7 @@ class Agent_Test_1(Ib_Agent):
         if len(self.state_history) > 10:
             self.state_history.pop()
 
-        at = np.array([0, 0, 0], dtype=np.float16)
+        at = np.array([0, 0, 0], dtype=np.float32)
         at[0] = action = (self.get_h('f', 0)-self.get_h('v', 3))
         # at[0] = ((-self.get_h('v', 3))-1.333333333333)
         # at[0] = ((-self.get_h('p', 9))-1.451234)
@@ -264,7 +264,7 @@ class Agent_Test_2(Ib_Agent):
         if len(self.state_history) > 10:
             self.state_history.pop()
 
-        at = np.array([0, 0, 0], dtype=np.float16)
+        at = np.array([0, 0, 0], dtype=np.float32)
         # at[0] = 12.31 * self.get_h('g', 8) - 11.079
         # at[0] = action = (self.get_h('f', 0)-self.get_h('v', 3))
         at[0] = ((-self.get_h('v', 3))-1.333333333333)
@@ -304,7 +304,7 @@ class Agent_Test_3(Ib_Agent):
         if len(self.state_history) > 10:
             self.state_history.pop()
 
-        at = np.array([0, 0, 0], dtype=np.float16)
+        at = np.array([0, 0, 0], dtype=np.float32)
         # at[0] = 12.31 * self.get_h('g', 8) - 11.079
         # at[0] = action = (self.get_h('f', 0)-self.get_h('v', 3))
         at[0] = ((-self.get_h('v', 3))-1.333333333333)
@@ -344,7 +344,7 @@ class Agent_Test_4(Ib_Agent):
         if len(self.state_history) > 10:
             self.state_history.pop()
 
-        at = np.array([0, 0, 0], dtype=np.float16)
+        at = np.array([0, 0, 0], dtype=np.float32)
         # at[0] = 12.31 * self.get_h('g', 8) - 11.079
         # at[0] = action = (self.get_h('f', 0)-self.get_h('v', 3))
         at[0] = ((-self.get_h('v', 3))-1.333333333333)
@@ -384,7 +384,7 @@ class Agent_Test_5(Ib_Agent):
         if len(self.state_history) > 10:
             self.state_history.pop()
 
-        at = np.array([0, 0, 0], dtype=np.float16)
+        at = np.array([0, 0, 0], dtype=np.float32)
         # at[0] = 12.31 * self.get_h('g', 8) - 11.079
         # at[0] = action = (self.get_h('f', 0)-self.get_h('v', 3))
         at[0] = ((-self.get_h('v', 3))-1.333333333333)
@@ -424,7 +424,7 @@ class Agent_Test_6(Ib_Agent):
         if len(self.state_history) > 10:
             self.state_history.pop()
 
-        at = np.array([0, 0, 0], dtype=np.float16)
+        at = np.array([0, 0, 0], dtype=np.float32)
         # at[0] = 12.31 * self.get_h('g', 8) - 11.079
         # at[0] = action = (self.get_h('f', 0)-self.get_h('v', 3))
         at[0] = ((-self.get_h('v', 3))-1.333333333333)
@@ -464,7 +464,7 @@ class Agent_Test_7(Ib_Agent):
         if len(self.state_history) > 10:
             self.state_history.pop()
 
-        at = np.array([0, 0, 0], dtype=np.float16)
+        at = np.array([0, 0, 0], dtype=np.float32)
         # at[0] = 12.31 * self.get_h('g', 8) - 11.079
         # at[0] = action = (self.get_h('f', 0)-self.get_h('v', 3))
         at[0] = ((-self.get_h('v', 3))-1.333333333333)
