+++ benchmarks/industrial_benchmark_python/ib_eval_agents.py
@@ -137,7 +137,7 @@ class Agent_daniel_21(Ib_Agent):
         if len(self.state_history) > 10:
             self.state_history.pop()
 
-        at = np.array([0, 0, 0], dtype=np.float16)
+        at = np.array([0, 0, 0], dtype=np.float32)
         at[0] = - self.get_h('v', 5) - 0.91
         at[1] = 2 * self.get_h('f', 3) - env_state['p'] + 1.43
         at[2] = -3.48 * self.get_h('h', 3) - self.get_h('h', 4) + 2*self.get_h('p', 0) + 0.81
@@ -156,7 +156,7 @@ class Agent_daniel_27(Ib_Agent):
         if len(self.state_history) > 10:
             self.state_history.pop()
 
-        at = np.array([0, 0, 0], dtype=np.float16)
+        at = np.array([0, 0, 0], dtype=np.float32)
         # # idk why, but these are the best??
         # at[0] = -self.get_h('v', 5) - 1.17
         # at[1] = 2 * self.get_h('f', 3) - self.get_h('p', 0) + 1.16
@@ -179,7 +179,7 @@ class Agent_Daniel_29_Best(Ib_Agent):
         if len(self.state_history) > 10:
             self.state_history.pop()
 
-        at = np.array([0, 0, 0], dtype=np.float16)
+        at = np.array([0, 0, 0], dtype=np.float32)
         at[0] = -self.get_h('v', 4) - 0.96  # 27
         at[1] = 0.41 * self.get_h('f', 1) + self.get_h('f', 4) - 0.59 * self.get_h('p', 0) + 0.77  # 27
         at[2] = -4.05 * self.get_h('h', 3) - self.get_h('h', 4) + 2.26 * self.get_h('p', 0) + 0.90  # 27
@@ -202,7 +202,7 @@ class Agent_Test(Ib_Agent):
         if len(self.state_history) > 10:
             self.state_history.pop()
 
-        at = np.array([0, 0, 0], dtype=np.float16)
+        at = np.array([0, 0, 0], dtype=np.float32)
         at[0] = -self.get_h('g', 5) - 1.8  # 27
         at[1] = 0.41 * self.get_h('f', 1) + self.get_h('f', 4) - 0.59 * self.get_h('p', 0) + 0.77  # -13.8*Consumption_30
         at[2] = -4.05 * self.get_h('h', 3) - self.get_h('h', 4) + 2.26 * self.get_h('p', 0) + 0.90  # 27
@@ -224,7 +224,7 @@ class Agent_Udluft(Ib_Agent):
         if len(self.state_history) > 10:
             self.state_history.pop()
 
-        at = np.array([0, 0, 0], dtype=np.float16)
+        at = np.array([0, 0, 0], dtype=np.float32)
         at[0] = -self.get_h('f', 0) - self.get_h('v', 3)  # 27
         at[1] = self.get_h('f', 4)
         at[2] = -6 * self.get_h('h', 4)
@@ -248,7 +248,7 @@ class Agent_Test(Ib_Agent):
         if len(self.state_history) > 10:
             self.state_history.pop()
 
-        at = np.array([0, 0, 0], dtype=np.float16)
+        at = np.array([0, 0, 0], dtype=np.float32)
         at[0] = 12.31*self.get_h('g', 8) - 11.079
         at[1] = -2.6/(self.get_h('c', 30) - self.get_h('f', 12) + self.get_h('f', 17))
         at[2] = 7.0442545165*self.get_h('p', 30) - 9.3079051182*self.get_h('h', 26) - 7.0442545165*self.get_h('h', 27) + 3.0
