+++ plagih/modules/tests.py
@@ -134,8 +134,8 @@ class MountainCarExamples:
     tree_test_plus_list = ['+', '-', '-', '*', '*', '*', '*', 1, 2, 3, 4, 5, 6, 7, 8]
     tree_test_plus_modify_v1 = [0, 0, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1]
 
-    fake_tensors = {'pos': tf.constant(1.1, dtype=tf.float16),
-                    'vel': tf.constant(2.2, dtype=tf.float16),
+    fake_tensors = {'pos': tf.constant(1.1, dtype=tf.float32),
+                    'vel': tf.constant(2.2, dtype=tf.float32),
                     'bl': tf.constant(True, dtype=tf.bool)}
 
     expr_test1 = 'Ifte(1.019*(-0.09)**b*(0.98 - 0.13) + Mini(b, pos) > -0.97, 0.0, 2.0)'
